requirements.txt
#Django==1.7c2
https://www.djangoproject.com/download/1.7c2/tarball/
Markdown==2.4.1
argparse==1.2.1
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==0.12
django-debug-toolbar==1.2.1
django-filter==0.7
django-redis==3.7.1
django-rest-swagger==0.1.14
django-rq==0.7.0
django-toolbelt==0.0.1
djangorestframework==2.4.2
fudge==1.0.3
google-api-python-client==1.2
gunicorn==19.1.1
psycopg2==2.5.4
python-memcached==1.53
python-openid==2.2.5
redis==2.10.3
requests==2.4.1
rq==0.4.6
simplejson==3.6.0
soundcloud==0.4.1
sqlparse==0.1.11
static3==0.5.1
wsgiref==0.1.2
